CerkeFS.Board2

Documentation

data Board1

data Error

Constructors

AlreadyOccupied Square
The square you're moving to is already occupied.

EmptySquare Square
The square you're moving from is actually empty

TamCapture
The square you're moving to is already occupied by Tam2.

NoCorrespondingPieceInHand
You tried to drop a piece that is not in the hand.

MovingOpponentPiece
You tried to move the opponent's piece.

FriendlyFire
You tried to capture your own piece.

AmbiguousColor
Color of the dropped piece cannot be unambiguously inferred.

WrongProfessionSpecified
The actual profession differs from the expectation.

Fields

expected :: Maybe Profession

specified :: Maybe Profession

FalseDeclaration
Declares a Dat2 whose condition is not satisfied.

Tam2HueAUai1Violation
Tried to take a piece protected by Tam2HueAUai1

SteppingEmptySquare Square
Tried to step on an empty square

ProfessionPrivilegeExceeded Profession Square
Trying a movement that the profession does not allow

CaptureByTam
Trying to capture a piece by Tam2

Tam2PrivilegeExceeded
Trying a movement that the Tam2 cannot do

Fields

_from :: Square

_thru :: Maybe Square

_to :: Square

sqKA :: Square

sqLA :: Square

sqNA :: Square

sqTA :: Square

sqZA :: Square

sqXA :: Square

sqCA :: Square

sqMA :: Square

sqPA :: Square

sqKE :: Square

sqLE :: Square

sqNE :: Square

sqTE :: Square

sqZE :: Square

sqXE :: Square

sqCE :: Square

sqME :: Square

sqPE :: Square

sqKI :: Square

sqLI :: Square

sqNI :: Square

sqTI :: Square

sqZI :: Square

sqXI :: Square

sqCI :: Square

sqMI :: Square

sqPI :: Square

sqKU :: Square

sqLU :: Square

sqNU :: Square

sqTU :: Square

sqZU :: Square

sqXU :: Square

sqCU :: Square

sqMU :: Square

sqPU :: Square

sqKO :: Square

sqLO :: Square

sqNO :: Square

sqTO :: Square

sqZO :: Square

sqXO :: Square

sqCO :: Square

sqMO :: Square

sqPO :: Square

sqKY :: Square

sqLY :: Square

sqNY :: Square

sqTY :: Square

sqZY :: Square

sqXY :: Square

sqCY :: Square

sqMY :: Square

sqPY :: Square

sqKAI :: Square

sqLAI :: Square

sqNAI :: Square

sqTAI :: Square

sqZAI :: Square

sqXAI :: Square

sqCAI :: Square

sqMAI :: Square

sqPAI :: Square

sqKAU :: Square

sqLAU :: Square

sqNAU :: Square

sqTAU :: Square

sqZAU :: Square

sqXAU :: Square

sqCAU :: Square

sqMAU :: Square

sqPAU :: Square

sqKIA :: Square

sqLIA :: Square

sqNIA :: Square

sqTIA :: Square

sqZIA :: Square

sqXIA :: Square

sqCIA :: Square

sqMIA :: Square

sqPIA :: Square

sqList :: [Square]

The list of squares, arranged in the following order: [sqKA, sqLA, sqNA ... sqCIA, sqMIA, sqPIA].

isOccupied :: Square -> Board1 -> Bool

Returns whether the square is occupied

data Square

type Square2 = Int

fromBoard1_old :: Board1_old -> Board1

toSquare :: Square2 -> Square

fromSquare :: Square -> Square2

isOccupied' :: Square2 -> Board1 -> Bool

lookup_ :: Square -> Board1 -> Maybe Piece