CerkeFS.GameState

Documentation

data Fullboard

Constructors

Fullboard

Fields

board :: Board1

hand :: [Piece]

plays :: Square -> Square -> Side -> Operation ()

Moves the piece. If the destination is blocked, the piece at the destination is implicitly captured.

MovingOpponentPiece is raised if the side of the moving piece does not match the side given in the argument.

FriendlyFire is raised if the destination is blocked by the piece belonging to the same side as the moving piece.

TamCapture is raised if the destination is occupied by Tam2, which cannot be captured.

FIXME: Uai protection not implemented

drops :: (Color, Profession) -> Square -> Side -> Operation ()

Drops the piece to the square.

NoCorrespondingPieceInHand is raised if no pieces in the hand matches the condition given by the arguments.

AlreadyOccupied is raised if the square is already occupied by another piece.

passes :: Side -> Operation ()

Skips a turn.

drops' :: Profession -> Square -> Side -> Operation ()

Similar to drops', but infers the color of the piece to be dropped. AmbiguousColor is raised if the color of the piece cannot be uniquely identified.

mun1 :: (Side -> Operation ()) -> Side -> Operation ()

Wraps an operation to show that the operation must theoretically succeed but did not happen. Fails if the wrapped operation is illegal.

plays' :: Square -> Profession -> Square -> Side -> Operation ()

Similar to plays, but also checks if the moving piece has the profession specified by the argument.

playsTam :: Square -> Square -> Side -> Operation ()

data Dat2

Constructors

Mun1MakMok1Hue
無抗行処; la als

Kua2Kauk2Mun1Aum1
筆兵無傾; la ny anknish

Huep2Hia1
地心; la meunerfergal

Mok1Mok1
行行; la nienulerless

Dat2AIo
王; la nermetixaler

Saup1
獣; la pysess

HuetKaikADat2
闇戦之集; la phertarsa'd elmss

Cuop2Mun1Mok1Hue
声無行処; la ytartanerfergal ~ | BapPok -- ^ 同色; la dejixece

declare :: Side -> Dat2 -> Operation ()

Declares a Dat2. Fails with FalseDeclaration if the condition required for the declaration is not met.

taxot1 :: Operation ()

type Operation = StateT Fullboard (Either Error)

(>+>) :: Monad m => (Side -> m a) -> (Side -> m b) -> m b

Upward plays, followed by Downward

(>->) :: Monad m => (Side -> m a) -> (Side -> m b) -> m b

Downward plays, followed by Upward