CerkeFS.Internal.Board

Documentation

data Col

Constructors

CK

CL

CN

CT

CZ

CX

CC

CM

CP

data Row

Constructors

RA

RE

RI

RU

RO

RY

RAI

RAU

RIA

data Square

Constructors

Square

Fields

row :: Row

col :: Col

type Board1 = Map Square Piece

data Vec

Constructors

Vec

Fields

dx :: Int

dy :: Int

putPiece :: Piece -> Square -> Board1 -> M Board1

Puts a piece on a square. Fails with AlreadyOccupied if already occupied.

removePiece :: Square -> Board1 -> M (Piece, Board1)

Removes a piece. Fails with EmptySquare if the specified square is empty.

movePieceFromToFull :: Square -> Square -> Board1 -> M (Maybe PhantomPiece, Board1)

Moves a piece on a square according to the vector. Raises:

OutOfBoard if the piece goes out of the board

EmptySquare if the original square is empty

AlreadyOccupied if the resulting square is already occupied

Moves a piece from a square to a square. Raises:

EmptySquare if the original square is empty

AlreadyOccupied if the resulting square is already occupied

data Error

Constructors

AlreadyOccupied Square
The square you're moving to is already occupied.

EmptySquare Square
The square you're moving from is actually empty

TamCapture
The square you're moving to is already occupied by Tam2.

NoCorrespondingPieceInHand
You tried to drop a piece that is not in the hand.

MovingOpponentPiece
You tried to move the opponent's piece.

FriendlyFire
You tried to capture your own piece.

AmbiguousColor
Color of the dropped piece cannot be unambiguously inferred.

WrongProfessionSpecified
The actual profession differs from the expectation.

Fields

expected :: Maybe Profession

specified :: Maybe Profession

FalseDeclaration
Declares a Dat2 whose condition is not satisfied.

Tam2HueAUai1Violation
Tried to take a piece protected by Tam2HueAUai1

add :: Vec -> Square -> Maybe Square

getNeighborsAndSelf :: Square -> [Square]

sqKA :: Square

sqLA :: Square

sqNA :: Square

sqTA :: Square

sqZA :: Square

sqXA :: Square

sqCA :: Square

sqMA :: Square

sqPA :: Square

sqKE :: Square

sqLE :: Square

sqNE :: Square

sqTE :: Square

sqZE :: Square

sqXE :: Square

sqCE :: Square

sqME :: Square

sqPE :: Square

sqKI :: Square

sqLI :: Square

sqNI :: Square

sqTI :: Square

sqZI :: Square

sqXI :: Square

sqCI :: Square

sqMI :: Square

sqPI :: Square

sqKU :: Square

sqLU :: Square

sqNU :: Square

sqTU :: Square

sqZU :: Square

sqXU :: Square

sqCU :: Square

sqMU :: Square

sqPU :: Square

sqKO :: Square

sqLO :: Square

sqNO :: Square

sqTO :: Square

sqZO :: Square

sqXO :: Square

sqCO :: Square

sqMO :: Square

sqPO :: Square

sqKY :: Square

sqLY :: Square

sqNY :: Square

sqTY :: Square

sqZY :: Square

sqXY :: Square

sqCY :: Square

sqMY :: Square

sqPY :: Square

sqKAI :: Square

sqLAI :: Square

sqNAI :: Square

sqTAI :: Square

sqZAI :: Square

sqXAI :: Square

sqCAI :: Square

sqMAI :: Square

sqPAI :: Square

sqKAU :: Square

sqLAU :: Square

sqNAU :: Square

sqTAU :: Square

sqZAU :: Square

sqXAU :: Square

sqCAU :: Square

sqMAU :: Square

sqPAU :: Square

sqKIA :: Square

sqLIA :: Square

sqNIA :: Square

sqTIA :: Square

sqZIA :: Square

sqXIA :: Square

sqCIA :: Square

sqMIA :: Square

sqPIA :: Square

sqList :: [Square]

The list of squares, arranged in the following order: [sqKA, sqLA, sqNA ... sqCIA, sqMIA, sqPIA].

isTam2HueAUai1 :: Side -> Board1 -> Square -> Bool