CerkeFS.Piece3

Documentation

data Color

Constructors

Kok1
Red, 赤

Huok2
Black, 黒

data Side

Constructors

Upward
Pieces that points upward. Denoted by ^ in the ASCII notation.

Downward
Pieces that points downward. Denoted by _ in the ASCII notation.

data Piece

data Profession

Constructors

Nuak1
Vessel, 船, felkana

Kauk2
Pawn, 兵, elmer

Gua2
Rook, 弓, gustuer

Kaun1
Bishop, 車, vadyrd

Dau2
Tiger, 虎, stistyst

Maun1
Horse, 馬, dodor

Kua2
Clerk, 筆, kua

Tuk2
Shaman, 巫, terlsk

Uai1
General, 将, varxle

Io
King, 王, ales

flipSide :: Piece -> Maybe Piece

Flip the side that the piece belongs to, or Nothing if Tam2.

type PhantomPiece

Arguments

= (Color, Profession, Side)
Not a physical piece; only serves to be a template

getSide :: Piece -> Maybe Side

Gets the side that the piece belongs to.

getSide' :: NonTam2Piece -> Side

toPhantom :: Piece -> Maybe PhantomPiece

Converts Piece into PhantomPiece.

船 :: Profession

An alias for Nuak1.

兵 :: Profession

An alias for Kauk2.

弓 :: Profession

An alias for Gua2.

車 :: Profession

An alias for Kaun1.

虎 :: Profession

An alias for Dau2.

馬 :: Profession

An alias for Maun1.

筆 :: Profession

An alias for Kua2.

巫 :: Profession

An alias for Tuk2.

将 :: Profession

An alias for Uai1.

王 :: Profession

An alias for Io.

黒 :: Color

An alias for Huok2.

赤 :: Color

An alias for Kok1.

phantomTam :: Maybe PhantomPiece

An alias for Nothing, that is, toPhantom Tam2.

data NonTam2Piece

toNonTam2Piece :: Piece -> Maybe NonTam2Piece

toFullPiece :: NonTam2Piece -> Piece

flipSide_ :: NonTam2Piece -> NonTam2Piece

Flip the side that the piece belongs to, or Nothing if Tam2.

toPhantom' :: NonTam2Piece -> PhantomPiece

Converts NonTam2Piece into PhantomPiece.