CerkeFS.PrettyPrint

Documentation

loadBoard :: String -> Maybe Board1

Converts the ASCII board format used in the Cerke board image generator to the Board1.

drawBoard :: Board1 -> String

Converts the Board1 to the ASCII board format used in the Cerke board image generator.