CerkeFS.VerifyRange

Documentation

vPlays2 :: Square -> Square -> Side -> Operation ()

Similar to plays, but checks whether the move is allowed by the profession.

vPlays3 :: Square -> Square -> Square -> Side -> Operation ()

Similar to plays, but checks whether the move is allowed by the profession.

The first arguments denotes the origin, the second denotes the square to step on, and the third denotes the destination.

vPlays3' :: Square -> Profession -> Square -> Square -> Side -> Operation ()

Similar to plays', but checks whether the move is allowed by the profession.

The first arguments denotes the origin, the third denotes the square to step on, and the fourth denotes the destination.

vPlays2' :: Square -> Profession -> Square -> Side -> Operation ()

Similar to plays', but checks whether the move is allowed by the profession.

vPlTam2 :: Square -> Square -> Side -> Operation ()

Plays Tam2, which did not step on a piece.

Thus, Tam2 must have visited an empty square, then to the destination.

vPlTam3 :: Square -> Square -> Square -> Side -> Operation ()

Plays Tam2, which stepped on a piece.

type Operation = StateT Fullboard (Either Error)

An alias for the monad that represents an operation.

taxot1 :: Operation ()

Supposed to be a declaration that the game has ended, but currently is just a dummy operation with no consequences.

drops :: (Color, Profession) -> Square -> Side -> Operation ()

Drops the piece to the square.

NoCorrespondingPieceInHand is raised if no pieces in the hand matches the condition given by the arguments.

AlreadyOccupied is raised if the square is already occupied by another piece.

drops' :: Profession -> Square -> Side -> Operation ()

Similar to drops', but infers the color of the piece to be dropped. AmbiguousColor is raised if the color of the piece cannot be uniquely identified.

mun1 :: (Side -> Operation a) -> Side -> Operation ()

Wraps an operation to show that the operation must theoretically succeed but did not happen. Fails if the wrapped operation is illegal.

declare :: Side -> Dat2 -> Operation ()

Declares a Dat2. Fails with FalseDeclaration if the condition required for the declaration is not met.

(>+>) :: Monad m => (Side -> m a) -> (Side -> m b) -> m b

Upward plays, followed by Downward

(>->) :: Monad m => (Side -> m a) -> (Side -> m b) -> m b

Downward plays, followed by Upward

data Dat2

Constructors

Mun1MakMok1Hue
無抗行処; la als

Kua2Kauk2Mun1Aum1
筆兵無傾; la ny anknish

Huep2Hia1
地心; la meunerfergal

Mok1Mok1
行行; la nienulerless

Dat2AIo
王; la nermetixaler

Saup1
獣; la pysess

HuetKaikADat2
闇戦之集; la phertarsa'd elmss

Cuop2Mun1Mok1Hue
声無行処; la ytartanerfergal