CerkeFS.VerifyRange

Documentation

vPlays2 :: Square -> Square -> Side -> Operation ()

Similar to plays, but checks whether the move is allowed by the profession.

vPlays3 :: Square -> Square -> Square -> Side -> Operation ()

Similar to plays, but checks whether the move is allowed by the profession.

| The first arguments denotes the origin, the second denotes the square to step on, and the third denotes the destination.

vPlays3' :: Square -> Profession -> Square -> Square -> Side -> Operation ()

Similar to plays', but checks whether the move is allowed by the profession.

| The first arguments denotes the origin, the third denotes the square to step on, and the fourth denotes the destination.

vPlays2' :: Square -> Profession -> Square -> Side -> Operation ()

Similar to plays', but checks whether the move is allowed by the profession.